+++ test/clj/cider/nrepl/middleware/slurp_test.clj
@@ -3,7 +3,8 @@
    [cider.nrepl.middleware.slurp :refer [slurp-url-to-content+body]]
    [clojure.java.io :as io]
    [clojure.test :as t]
-   [clojure.string :as str]))
+   [clojure.string :as str]
+   [orchard.misc :refer [java-api-version]]))
 
 (t/deftest test-project-clj-is-clj
   (let [resp (-> "project.clj"
@@ -14,12 +15,13 @@
     (t/is (= ["text/clojure" {}] (:content-type resp)))
     (t/is (not= "base64" (:content-transfer-encoding resp)))))
 
-(t/deftest test-sum-types-is-base64
-  (let [resp (slurp-url-to-content+body
-              (.toString
-               (io/resource "sum-types-are-cool.jpg")))]
-    (t/is (= ["image/jpeg" {}] (:content-type resp)))
-    (t/is (= "base64" (:content-transfer-encoding resp)))))
+(when-not (= 8 java-api-version)
+  (t/deftest test-sum-types-is-base64
+    (let [resp (slurp-url-to-content+body
+                (.toString
+                 (io/resource "sum-types-are-cool.jpg")))]
+      (t/is (= ["image/jpeg" {}] (:content-type resp)))
+      (t/is (= "base64" (:content-transfer-encoding resp))))))
 
 (t/deftest test-unrecognized-file
   (let [resp (slurp-url-to-content+body