+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Bump `orchard` to [0.22.0](https://github.com/clojure-emacs/orchard/blob/v0.22.0/CHANGELOG.md#0220-2024-01-14).
 * Bump `logjam` to [0.2.0](https://github.com/clojure-emacs/logjam/blob/v0.2.0/CHANGELOG.md#020-2024-01-04).
 * Bump `suitable` to [0.6.2](https://github.com/clojure-emacs/clj-suitable/blob/v0.6.2/CHANGELOG.md#062-2033-01-14).
+* Bump `compliment` to [0.5.1](https://github.com/alexander-yakushev/compliment/blob/6119d8/CHANGELOG.md#051-2023-11-30).
 * [#842](https://github.com/clojure-emacs/cider-nrepl/issues/842): Initial cache warmup: ensure that no classes' static initializers are run.
 * Forcibly exit the JVM on older (unsupported) Clojure versions.
   * This can help users and maintainers alike diagnose issues more quickly, avoiding problematic code paths in our middleware, and clients like cider.el.